Using Unity Version of AdColony SDK
The Unity Version only includes armeabi-v7a and x86 versions of the libraries. The non-unity version has 64-bit versions of the libraries as well.

When you run the application on a device with an arm64-v8a processor with the full library, it finds the two 64-bit adcolony libraries, and then crashes, for it expects to find 64-bit versions of all the other libraries (such as the Unity and Mono libraries) and they don't exist.

Using the Unity version of the libraries, the processor loads the older architecture's library, finds matching libraries for other modules, and works fine.